==> 07_metrics_across_backends.Rmd <==
---
title: "Omnibenchmark backend comparison"
date: "`r format(Sys.time(), '%d %B, %Y %H:%M:%S')`"
output:
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
    code_download: true
    number_sections: true
    df_print: kable
    theme: lumen
params:
    performance_bn: "performance-results.rds" # these are ignored in this report
    metrics_bn: "metrics-results.rds"         # the metrics basenames as generated by `04_metric_collector.Rmd`
    clustering_dir: ""                        # the path containing all `out` ob folders
editor_options: 
  chunk_output_type: console
---


```{r}
library(dplyr)
library(ggplot2)
library(knitr)
```

`bn` stands for basename (e.g. the RDS basenames generated by `04_metric_collector.Rmd`) and `clustering_dir` stands for the parent directory with all `out_conda`, `out_singularity` etc dirs. 

This report aims to generalize https://github.com/imallona/clustering_report/blob/6b4765d081f8882c7f7f40bf85fddefc02c1c7f2/quick-dirty-analysis.R

```
rmarkdown::render("07_metrics_across_backends.Rmd", params = list(
  performance_bn = "performance-results.rds",
  metrics_bn = "metrics-results.rds",
  clustering_dir =  "/home/imallona/src/clustering_example"))
```

```{r}
opts_chunk$set(fig.width = 6,
               fig.height = 6,
               cache = FALSE,
               include = TRUE,
               dev = "png",
               cache.lazy = FALSE,
               warning = TRUE,
               message = TRUE)
```

```{r}
## this will get as many performance/metrics files as runs there were
## params <- list(performance_bn = "performance-results.rds",
##                metrics_bn = "metrics-results.rds",
##                clustering_dir =  "/home/imallona/src/clustering_example")

(runs <- list.files(pattern = params$metrics_bn, path = params$clustering_dir, recursive = TRUE))
## runs <- runs[2:5]
```


# Pairwise compare {.tabset .tabset-pills}

```{r, results = 'asis'}

res <- list()

for (run1 in runs) {
    cat('## ', run1, '{.tabset .tabset-pills} \n\n')

    r1 <- readRDS(file.path(params$clustering_dir, run1))
    r1.id <- dirname(paste0(".", run1))
    
    for (run2 in setdiff(runs, run1)) {
        
        cat('### ', run2, '\n\n')
    
        r2 <- readRDS(file.path(params$clustering_dir, run2))
        r2.id <- dirname(paste0(".", run2))
        
        df <- r1 %>% select(data, method, metric, k) %>% 
            full_join(r2 %>% select(data, method, metric, k),
                      by = c("data", "method", "metric"), 
                      suffix = c(r1.id, r2.id))
        
        print(ggplot(df, aes(x=.data[[paste0('k', r1.id)]],
                             y=.data[[paste0('k', r2.id)]],
                             colour = data)) +
              geom_point(size = 3) +
              geom_abline() +
              theme(legend.position = "none") +
              facet_wrap(~metric, scales = "free") +
              theme(aspect.ratio=1))

        print( ggplot(df, aes(x=.data[[paste0('k', r1.id)]],
                              y=.data[[paste0('k', r2.id)]],
                              colour = data,
                              fill = method)) +
               geom_point(size = 3) +
               geom_abline() +
               theme(legend.position = "none") +
               facet_wrap(~method, scales = "free") +
               theme(aspect.ratio=1))

        ## print(ggplot(df,
        ##              aes(y=.data[[paste0('k', r1.id)]] - .data[[paste0('k', r2.id)]],
        ##                  x=method,
        ##                  colour = method)) +
        ##       geom_point(size = 3) +
        ##       geom_abline() +
        ##       theme(legend.position = "bottom") +
        ##       facet_wrap(~metric, scales = "free") +
        ##       theme(aspect.ratio=1))

        ## print(ggplot(df, aes(y=.data[[paste0('k', r1.id)]] - .data[[paste0('k', r2.id)]],
        ##                      x=data,
        ##                      colour = data)) +
        ##       geom_point(size = 3) +
        ##       geom_abline() +
        ##       theme(legend.position = "bottom") +
        ##       facet_wrap(~metric, scales = "free") +
        ##       theme(aspect.ratio=1))

        res[[paste(r1.id, r2.id)]] <- c(r1.id, r2.id, cor(r1$k, r2$k))

        cat('\n\n')
    }

    cat('### Internal check\n')
    
    ## additional check (idea from Ben): are repeated ks consistent within method:dataset runs?
    ## fcps-atom has a cardinality of 2, so k-1, k-2 etc should match true k (metrics results)
    atom <- r1[r1$data == 'fcps|atom',]
    atom$identical_true_m1 <- atom$k == atom$`k-1`
    atom$identical_true_m2 <- atom$k == atom$`k-2`
    odd <- atom[(!atom$identical_true_m1 | !atom$identical_true_m2),]
    if (nrow(odd) > 0) {
        cat('inconsistent indeed\n\n')
        print(kable(odd[,c('method', 'metric', 'k-2', 'k-1', 'k', 'identical_true_m1', 'identical_true_m2')], "html"))
    } else {
        print('redundant ks consistency check passed - identical for atom')
    }
}

```

# Export

Export correlations at true `k`.

```{r}
res <- do.call(rbind.data.frame, res)
colnames(res) <- c('id1', 'id2', 'cor_at_true_k')
res$cor_at_true_k <- as.numeric(res$cor_at_true_k)

print(res)

write.csv(file = '07_metrics_across_backends__metrics_correlation.csv', res, row.names = FALSE)

```
